requirements.txt
absl-py==2.0.0
aiofile==3.8.8
aiofiles==23.2.1
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiohttp-retry==2.8.3
aiohttpx==0.0.11
aiosignal==1.3.1
aiosqlite==0.20.0
amqp==5.2.0
annotated-types==0.6.0
anthropic==0.25.6
anyascii==0.3.2
anyio==3.7.1
APScheduler==3.10.4
asttokens==2.4.1
async-lru==2.0.4
async-openai==0.0.40
attrs==23.1.0
audioread==3.0.1
Babel==2.13.1
backoff==2.2.1
bangla==0.0.2
bcrypt==4.1.2
beautifulsoup4==4.12.2
billiard==4.2.0
blinker==1.7.0
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.6
boto3==1.34.155
botocore==1.34.155
cachetools==5.3.2
caio==0.9.13
catalogue==2.0.10
celery==5.4.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.16.0
cohere==5.6.2
colorama==0.4.6
colorlog==6.8.2
confection==0.1.4
contourpy==1.2.0
coqpit==0.0.17
cryptography==42.0.8
cycler==0.12.1
cymem==2.0.8
Cython==3.0.6
dataclasses-json==0.6.6
dateparser==1.1.8
decorator==5.1.1
deepgram-sdk==3.2.7
defusedxml==0.7.1
dill==0.3.7
distro==1.8.0
dnspython==2.6.1
docopt==0.6.2
dramatiq==1.17.0
dramatiq-dashboard==0.4.0
ecdsa==0.19.0
einops==0.7.0
elevenlabs==0.2.27
emoji==2.9.0
encodec==0.1.1
eventlet==0.35.2
executing==2.0.1
fastapi==0.104.1
fastavro==1.9.5
ffmpeg-python==0.2.0
file-io==0.5.0
filelock==3.13.1
Flask==3.0.0
Flask-Login==0.6.3
Flask-Sockets==0.2.1
fonttools==4.46.0
fpdf2==2.7.8
freetype-py==2.4.0
frozendict==2.3.10
frozenlist==1.4.0
fsspec==2023.12.0
future==1.0.0
g2pkk==0.1.2
gevent==23.9.1
gevent-websocket==0.10.1
google==3.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.14.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
google-cloud==0.34.0
google-cloud-texttospeech==2.14.2
google-generativeai==0.8.2
google-genai>=1.32.0
googleapis-common-protos==1.61.0
greenlet==3.0.2
grpcio==1.59.3
grpcio-status==1.59.3
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
gunicorn==22.0.0
h11==0.14.0
hangul-romanize==0.1.0
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.25.2
httpx-sse==0.4.0
huggingface-hub==0.19.4
idna==3.6
inflect==7.0.0
ipython==8.18.1
itsdangerous==2.1.2
jamo==0.4.1
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.2
jiter==0.5.0
jmespath==1.0.1
joblib==1.3.2
jsonlines==1.2.0
keyboard==0.13.5
kiwisolver==1.4.5
kombu==5.4.1
langcodes==3.3.0
lazy_loader==0.3
lazyops==0.2.61
librosa==0.10.1
llvmlite==0.41.1
loguru==0.7.2
Markdown==3.5.1
markdown2==2.4.13
MarkupSafe==2.1.3
marshmallow==3.21.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
murmurhash==1.0.10
mutagen==1.47.0
mypy-extensions==1.0.0
networkx==2.8.8
nltk==3.8.1
num2words==0.5.13
numba==0.58.1
numpy==1.26.2
oauthlib==3.2.2
openai==1.40.0
orjson==3.11.6
packaging==23.2
pandas==1.5.3
parameterized==0.9.0
parso==0.8.3
passlib==1.7.4
pillow==10.3.0
platformdirs==4.0.0
pooch==1.8.0
preshed==3.0.9
prometheus_client==0.21.0
prompt-toolkit==3.0.43
proto-plus==1.22.3
protobuf==4.23.4
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.8.2
pydantic-settings==2.1.0
pydantic_core==2.20.1
pydub==0.25.1
pygame==2.5.2
Pygments==2.17.2
PyJWT==2.8.0
pynndescent==0.5.11
PyOpenGL==3.1.7
pyparsing==3.1.1
pyperclip==1.8.2
pypinyin==0.49.0
pypng==0.20220715.0
pysbd==0.3.4
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
# MIGRATED TO PyJWT: python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==26.2.0
qrcode==7.4.2
redis==4.6.0
regex==2024.7.24
reportlab==4.2.0
requests==2.31.0
requests-mock==1.12.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
semantic-router==0.0.55
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
socketify==0.0.27
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
srsly==2.4.8
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.27.0
starlette-session==0.4.3
SudachiDict-core==20230927
SudachiPy==0.6.7
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
thinc==8.2.1
threadpoolctl==3.2.0
tiktoken==0.7.0
tokenizers==0.15.0
torch==2.1.1
torchaudio==2.1.1
tqdm==4.66.1
trainer==0.0.32
traitlets==5.14.0
transformers==4.35.2
TTS==0.21.3
twilio==8.10.3
typer==0.9.0
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.1
tzdata==2023.3
tzlocal==5.2
umap-learn==0.5.5
Unidecode==1.3.7
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.24.0.post1
verboselogs==1.7
vine==5.1.0
wasabi==1.1.2
watchfiles==0.22.0
Wave==0.0.2
wcwidth==0.2.12
weasel==0.3.4
websockets==12.0
Werkzeug==3.0.1
wfastcgi==3.0.0
win32-setctime==1.1.0 ; sys_platform == 'win32'
yarl==1.9.3
zope.event==5.0
zope.interface==6.1
stripe>=7.0.0
